rm sign/unsign warning
Former-commit-id: 1894eaa5e270630835d68055a99b3811563431cf

+++ include/igl/map_vertices_to_circle.cpp
@@ -26,7 +26,7 @@ IGL_INLINE void igl::map_vertices_to_circle(
     map_ij[bnd[i]] = i;
   }
 
-  for (int i = 0; i < isOnBnd.size(); i++)
+  for (int i = 0; i < (int)isOnBnd.size(); i++)
   {
     if (!isOnBnd[i])
     {
